# infrastructure/cdk.out/asset.6e06e72e0d4f224c8d22cdbe117d9bfc30a15192777cb5c986e4382000968dc3/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Hlekkr Media Processing API
  description: |
    Comprehensive API for media upload, deepfake detection, trust score calculation, 
    source verification, chain of custody tracking, and discrepancy detection.
    
    ## Authentication
    All endpoints (except health check) require JWT authentication via the Authorization header:
    ```
    Authorization: Bearer <jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited per user. Rate limit headers are included in responses.
    
    ## Error Handling
    All errors follow a consistent format with correlation IDs for tracking.
  version: 1.0.0
  contact:
    name: Hlekkr API Support
    email: api-support@hlekkr.com
  license:
    name: Proprietary
servers:
  - url: https://api.hlekkr.com/v1
    description: Production server
  - url: https://staging-api.hlekkr.com/v1
    description: Staging server
  - url: https://dev-api.hlekkr.com/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /trust-scores:
    get:
      summary: Get Trust Scores
      description: Retrieve trust scores with optional filtering and statistics
      operationId: getTrustScores
      tags:
        - Trust Scores
      parameters:
        - name: scoreRange
          in: query
          description: Filter by score range
          schema:
            type: string
            enum: ['0-20', '21-40', '41-60', '61-80', '81-100']
        - name: startDate
          in: query
          description: Start date for filtering (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for filtering (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: minScore
          in: query
          description: Minimum trust score
          schema:
            type: number
            minimum: 0
            maximum: 100
        - name: maxScore
          in: query
          description: Maximum trust score
          schema:
            type: number
            minimum: 0
            maximum: 100
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: statistics
          in: query
          description: Include statistical summary
          schema:
            type: boolean
            default: false
        - name: days
          in: query
          description: Number of days to look back
          schema:
            type: integer
            minimum: 1
            maximum: 365
      responses:
        '200':
          description: Trust scores retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustScoresListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /trust-scores/{mediaId}:
    get:
      summary: Get Media Trust Score
      description: Retrieve trust score for a specific media item
      operationId: getMediaTrustScore
      tags:
        - Trust Scores
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
        - name: history
          in: query
          description: Include score history
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          description: Maximum number of historical records
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Trust score retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustScoreResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    
    post:
      summary: Calculate Trust Score
      description: Calculate or recalculate trust score for a media item
      operationId: calculateTrustScore
      tags:
        - Trust Scores
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        description: Trust score calculation options
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateTrustScoreRequest'
      responses:
        '200':
          description: Trust score calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustScoreResponse'
        '202':
          description: Trust score calculation started (async)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /source-verification/{mediaId}:
    post:
      summary: Verify Media Source
      description: Perform source verification for a media item
      operationId: verifySource
      tags:
        - Source Verification
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        description: Source verification request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifySourceRequest'
      responses:
        '200':
          description: Source verification completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceVerificationResponse'
        '202':
          description: Source verification started (async)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    
    get:
      summary: Get Source Verification Results
      description: Retrieve source verification results for a media item
      operationId: getSourceVerification
      tags:
        - Source Verification
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Source verification results retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceVerificationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /chain-of-custody:
    post:
      summary: Record Custody Event
      description: Record a new event in the chain of custody
      operationId: recordCustodyEvent
      tags:
        - Chain of Custody
      requestBody:
        required: true
        description: Custody event details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordCustodyEventRequest'
      responses:
        '201':
          description: Custody event recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustodyEventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /chain-of-custody/{mediaId}:
    get:
      summary: Get Custody Chain
      description: Retrieve the complete chain of custody for a media item
      operationId: getCustodyChain
      tags:
        - Chain of Custody
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
        - name: operation
          in: query
          description: Filter by operation type
          schema:
            type: string
            enum: ['upload', 'analysis', 'review', 'approval', 'rejection', 'modification', 'deletion']
      responses:
        '200':
          description: Custody chain retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustodyChainResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /chain-of-custody/{mediaId}/provenance:
    get:
      summary: Get Provenance Visualization Data
      description: Get data for visualizing media provenance and custody chain
      operationId: getProvenance
      tags:
        - Chain of Custody
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Provenance data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvenanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /chain-of-custody/{mediaId}/verify:
    get:
      summary: Verify Chain Integrity
      description: Verify the cryptographic integrity of the custody chain
      operationId: verifyChainIntegrity
      tags:
        - Chain of Custody
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chain integrity verification completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainIntegrityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /discrepancies:
    post:
      summary: Detect Discrepancies
      description: Analyze system data for discrepancies and anomalies
      operationId: detectDiscrepancies
      tags:
        - Discrepancy Detection
      requestBody:
        description: Discrepancy detection parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectDiscrepanciesRequest'
      responses:
        '200':
          description: Discrepancy analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscrepancyResponse'
        '202':
          description: Discrepancy analysis started (async)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    
    get:
      summary: Get Recent Discrepancies
      description: Retrieve recently detected discrepancies with filtering
      operationId: getDiscrepancies
      tags:
        - Discrepancy Detection
      parameters:
        - name: timeRangeHours
          in: query
          description: Time range in hours to look back
          schema:
            type: integer
            minimum: 1
            maximum: 8760
            default: 24
        - name: severityThreshold
          in: query
          description: Minimum severity level
          schema:
            type: string
            enum: ['low', 'medium', 'high', 'critical']
            default: 'medium'
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Discrepancies retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscrepancyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /discrepancies/{mediaId}:
    post:
      summary: Analyze Media Discrepancies
      description: Analyze a specific media item for discrepancies
      operationId: analyzeMediaDiscrepancies
      tags:
        - Discrepancy Detection
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media discrepancy analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscrepancyResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    
    get:
      summary: Get Media Discrepancies
      description: Retrieve discrepancies for a specific media item
      operationId: getMediaDiscrepancies
      tags:
        - Discrepancy Detection
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media discrepancies retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscrepancyResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /discrepancies/patterns:
    post:
      summary: Analyze Suspicious Patterns
      description: Analyze system-wide patterns for suspicious activity
      operationId: analyzePatterns
      tags:
        - Discrepancy Detection
      parameters:
        - name: timeRangeHours
          in: query
          description: Time range in hours to analyze
          schema:
            type: integer
            minimum: 1
            maximum: 8760
            default: 168
        - name: minSeverity
          in: query
          description: Minimum severity for pattern detection
          schema:
            type: string
            enum: ['low', 'medium', 'high', 'critical']
            default: 'medium'
      responses:
        '200':
          description: Pattern analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatternAnalysisResponse'
        '202':
          description: Pattern analysis started (async)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /media:
    post:
      summary: Upload Media
      description: Create a presigned URL for media upload
      operationId: uploadMedia
      tags:
        - Media Upload
      requestBody:
        required: true
        description: Media upload request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadMediaRequest'
      responses:
        '200':
          description: Upload URL created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadMediaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
  /media/{mediaId}:
    get:
      summary: Get Media Information
      description: Retrieve information about a media item
      operationId: getMedia
      tags:
        - Media Upload
      parameters:
        - name: mediaId
          in: path
          required: true
          description: Media identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaInfoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication service
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    TrustScoresListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                trustScores:
                  type: array
                  items:
                    $ref: '#/components/schemas/TrustScore'
                pagination:
                  $ref: '#/components/schemas/Pagination'
                statistics:
                  $ref: '#/components/schemas/TrustScoreStatistics'
...